requirements.txt
crewai>=0.80.0
crewai-tools>=0.12.0
python-dotenv>=1.0.1
requests>=2.31.0

# File parsing
pypdf>=4.2.0
python-docx>=1.1.2

# Web UI
streamlit>=1.36.0

# API Server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# PDF Generation
reportlab>=4.0.0
fpdf2>=2.7.0

# Utilities
pydantic>=2.8.2
